change upload to not ignore errors
the repo file adding api is entirely awkward as it avoids HTTP codes
altogether and returns errorish types even when there is none.
this requires additional work on our end and in particular error
representation is distinct from other errors.

to sensibly behave here add_file will return only the Added part of the
file operation report. in the event of an error a RepositoryFileError
is raised which is to_s'ing further information about the error cause.
unfortunately the cause format itself is not entirely specified so
from a user POV this is either an all or nothing sort of deal right now.