make sure publish cannot be called with clearly invalid prefixes
prefixes passed into the publish REST api are encoded as /=>_ and _=>__
due to the nature of this encoding we cannot automatically convert this
without breaking backwards compatibility foo_bar may be meant to be foo/bar
or actual foo_bar.
raise an invalidity error when a slash is detected, when a _ is detected
the prefix is valid, but may end up misformed if foo__bar was meant.
somewhat unfortunate but not much to be done about this on our end.

with next major version we should break this and turn it inside out such
that the expected behavior is to pass in
foo/bar and we'll autoconvert it into api-safe form.