add new Files.tmp_upload wrapper, use it in Repo.upload
this makes the temporary upload ability accessible from the outside so
that one can upload a set of files and then in a block add them to multiple
different repositories. previously one would have had to reimplement
Repo.upload which might use a cheaper tmpdir implementation which
would run risk of conflicting on the remote.
with this new method you get all the usefulness of Repo.upload without
it's single-reponess